backend/requirements.txt
# SignalCraft Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Analysis
duckdb
pandas
pyarrow
scipy

# Broker Integrations & External Clients
requests
websockets
yfinance
pyotp
kiteconnect
NorenRestApiPy
python-multipart
redis==5.0.3
